ospf6d: port LLS correction from ospfd
It was recently spotted in ospf_packet_examin(), that an OSPFv2 packet
with trailing LLS data has its "length" OSPF header field value less,
than the amount of bytes on wire (which is OK considering both specs).
This is also true for OSPFv3. It is hard (if ever possible) to find an
LLS-enabled OSPFv3 implementation today, this is why a wrong check
condition in ospf6_packet_examin() did not break things thus far. This
commit fixes the mistake.