lib: introduce ip_masklen_safe()
ip_masklen() function is used to convert a 32-bit long netmask to a
[0..32] masklen. The last optimisation made it sometimes return negative
signed result in case of erroneous (non-contiguous) input. This signed
value, converted to unsigned type, stands for a value much larger than
32, which in turn triggers an assertion failure in apply_mask_ipv4(),
which often follows a call to ip_masklen(). This commit adds a new
function, which can reliably detect errors in the netmask, and makes
netmask_str2prefix_str() use it.