ospfd: fix iterations over multiple-TOS links
Some of the functions iterating over link descriptor blocks of a
Router-LSA used to do it assuming all links being single-TOS (12 bytes
long). Several others did it right, deriving the offset from "# TOS"
(router_lsa_link.m[0].tos_count) field. This commit fixes a few
functions in this regard. It should be noted, that additional TOS
metrics are only meaningful in the scope of offset calculation and are
ignored elsewhere.

* ospf_dump.c (ospf_router_lsa_dump)
* ospf_packet.c (ospf_router_lsa_links_examin, ospf_lsa_netmasks_examin)
* ospf_spf.c (ospf_lsa_has_link)
* ospf_vty.c (show_ip_ospf_database_router_links)