bgpd: fix error code for OPEN errors (RFC4271 6.2)
"All errors detected while processing the OPEN message MUST be
indicated by sending the NOTIFICATION message with the Error Code
OPEN Message Error.  The Error Subcode elaborates on the specific
nature of the error.
[...]
If one of the Optional Parameters in the OPEN message is recognized,
but is malformed, then the Error Subcode MUST be set to 0
(Unspecific)."

* bgp_open.c
  * bgp_capability_orf_entry()
  * bgp_capability_parse()
  * bgp_open_option_parse()