ripd: implement RFC4822 authentication key compression
Recent justifications in authentication keys handling belong to a
solution of a bigger problem, which this change should make complete.

RIP-2 cryptographic authentication implemented in ripd so far used the
original HMAC construct for message digest generation. However, RFC4822
construct is different (see the comment block added by this commit).
That said, it is possible to implement the RFC4822 construct on top of
the HMAC construct.

This change adds a key compression function to make ripd cryptographic
(except Keyed MD5) authentication compliant to RFC4822 without losing
the convenience of Libgcrypt.

* lib/cryptohash.c
  * hash_key_compress_rfc4822(): new function
* ripd/rip_auth.c
  * rip_auth_check_hash(): update to use hash_key_compress_rfc4822()
  * rip_auth_write_trailer(): idem