bgpd: Treat bnc resolved nexthop list as linked list
During testing I saw strange behavior of "sh ip bgp scan detail". In
case when some recursive nexthop resolved to multipath route, the
output was:

BGP scan is running
BGP scan interval is 5
Current BGP nexthop cache:
 10.100.255.1 valid [IGP metric 20]
  gate 192.168.0.3
  invalid nexthop type 61
BGP connected route:
 192.168.0.0/24
 192.168.1.0/24

(The "type" was changing from time to time)

Whereas correct output should be:

...
  gate 192.168.0.3
  gate 192.168.1.3
...

The reason for this behavior is that bnc_nexthop_cache "nexthop"
member was erroneously treated as array, not as linked list as it
should be.

This patch corrects the issue.